Move test case from TypeServiceImplIT to TypeServiceImplTest
